git-svn-id: https://svn.apache.org/repos/asf/lucene/hadoop/branches/branch-0.13@543623 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/hadoop/mapred/TaskTracker.java
@@ -1930,7 +1930,7 @@ public class TaskTracker
         //seek to the correct offset for the reduce
         mapOutputIn.seek(startOffset);
           
-        int totalRead = 0;
+        long totalRead = 0;
         int len = mapOutputIn.read(buffer, 0,
                                    partLength < MAX_BYTES_TO_READ 
                                    ? (int)partLength : MAX_BYTES_TO_READ);
